Create a few initial subdirectories in the installed dir to mimic a standard install
We've seen a crash when you install an app through Sim Deploy that didn't happen in a standard
Xcode build because the tmp subdirectory didn't exist in the Sim Deploy version.  Granted,
it's bad form to make assumptions about directories existing, but it seems like the
best thing to do is try to make a SD install as close to an Xcode install as possible.